// packages/brand/next/head-metadata/from-next/opengraph-types.ts
export type OpenGraph =
  | OpenGraphWebsite
  | OpenGraphArticle
  | OpenGraphBook
  | OpenGraphProfile
  | OpenGraphMusicSong
  | OpenGraphMusicAlbum
  | OpenGraphMusicPlaylist
  | OpenGraphRadioStation
  | OpenGraphVideoMovie
  | OpenGraphVideoEpisode
  | OpenGraphVideoTVShow
  | OpenGraphVideoOther
  | OpenGraphMetadata

// update this type to reflect actual locales
type Locale = string

type OpenGraphMetadata = {
  determiner?: 'a' | 'an' | 'the' | 'auto' | ''
  title?: string | TemplateString
  description?: string
  emails?: string | Array<string>
  phoneNumbers?: string | Array<string>
  faxNumbers?: string | Array<string>
  siteName?: string
  locale?: Locale
  alternateLocale?: Locale | Array<Locale>
  images?: OGImage | Array<OGImage>
  audio?: OGAudio | Array<OGAudio>
  videos?: OGVideo | Array<OGVideo>
  url?: string | URL
  countryName?: string
  ttl?: number
}
export type OpenGraphWebsite = OpenGraphMetadata & {
  type: 'website'
}
type OpenGraphArticle = OpenGraphMetadata & {
  type: 'article'
  publishedTime?: string // datetime
  modifiedTime?: string // datetime
  expirationTime?: string // datetime
  authors?: null | string | URL | Array<string | URL>
  section?: null | string
  tags?: null | string | Array<string>
}
type OpenGraphBook = OpenGraphMetadata & {
  type: 'book'
  isbn?: null | string
  releaseDate?: null | string // datetime
  authors?: null | string | URL | Array<string | URL>
  tags?: null | string | Array<string>
}
type OpenGraphProfile = OpenGraphMetadata & {
  type: 'profile'
  firstName?: null | string
  lastName?: null | string
  username?: null | string
  gender?: null | string
}
type OpenGraphMusicSong = OpenGraphMetadata & {
  type: 'music.song'
  duration?: null | number
  albums?: null | string | URL | OGAlbum | Array<string | URL | OGAlbum>
  musicians?: null | string | URL | Array<string | URL>
}
type OpenGraphMusicAlbum = OpenGraphMetadata & {
  type: 'music.album'
  songs?: null | string | URL | OGSong | Array<string | URL | OGSong>
  musicians?: null | string | URL | Array<string | URL>
  releaseDate?: null | string // datetime
}
type OpenGraphMusicPlaylist = OpenGraphMetadata & {
  type: 'music.playlist'
  songs?: null | string | URL | OGSong | Array<string | URL | OGSong>
  creators?: null | string | URL | Array<string | URL>
}
type OpenGraphRadioStation = OpenGraphMetadata & {
  type: 'music.radio_station'
  creators?: null | string | URL | Array<string | URL>
}
type OpenGraphVideoMovie = OpenGraphMetadata & {
  type: 'video.movie'
  actors?: null | string | URL | OGActor | Array<string | URL | OGActor>
  directors?: null | string | URL | Array<string | URL>
  writers?: null | string | URL | Array<string | URL>
  duration?: null | number
  releaseDate?: null | string // datetime
  tags?: null | string | Array<string>
}
type OpenGraphVideoEpisode = OpenGraphMetadata & {
  type: 'video.episode'
  actors?: null | string | URL | OGActor | Array<string | URL | OGActor>
  directors?: null | string | URL | Array<string | URL>
  writers?: null | string | URL | Array<string | URL>
  duration?: null | number
  releaseDate?: null | string // datetime
  tags?: null | string | Array<string>
  series?: null | string | URL
}
type OpenGraphVideoTVShow = OpenGraphMetadata & {
  type: 'video.tv_show'
}
type OpenGraphVideoOther = OpenGraphMetadata & {
  type: 'video.other'
}

export type OGImage = string | OGImageDescriptor | URL
type OGImageDescriptor = {
  url: string | URL
  secureUrl?: string | URL
  alt?: string
  type?: string
  width?: string | number
  height?: string | number
}
type OGAudio = string | OGAudioDescriptor | URL
type OGAudioDescriptor = {
  url: string | URL
  secureUrl?: string | URL
  type?: string
}
type OGVideo = string | OGVideoDescriptor | URL
// ...
type OGSong = {
  url: string | URL
  disc?: number
  track?: number
}
type OGAlbum = {
  url: string | URL
  disc?: number
  track?: number
}
type OGActor = {
  url: string | URL
  role?: string
}